Bug #1380275 pclint: PClint and remove some logs
[root cause  ] modify as pclint suggestion
[changes     ] hal3_2v6,oem2v6
[side effects] no
[self test   ]: yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]: ums512_1h10
[test case]:camera normal function
[reviewers   ] Victor.Xu,Alwin.Zou
[change_type ] bugfix
[tag_product ] common

+++ oem2v6/src/SprdOEMCamera.c
@@ -1204,7 +1204,7 @@ cmr_int camera_ioctrl(cmr_handle handle, int cmd, void *param) {
         struct camera_context *cxt = (struct camera_context *)handle;
 
         ret = cmr_sensor_get_stream_status(cxt->sn_cxt.sensor_handle, cxt->camera_id);
-        *((int *)param) = ret;
+        *((int *)param) = (int)ret;
         break;
     }
     case CAMERA_TOCTRL_SET_HIGH_RES_MODE: {
